agregado de produtos comprados (no terminado)

+++ Api/src/handlers/usersHanlders.js
@@ -1,4 +1,4 @@
-const { getAllWishes,getAllUsers,addUser,addProductInShoppingCartForUser,addFriends,addWishToList,getAllFriends,getAllProductsInShoppingCart,addNewComment,getAllCommentOfUser,getAllCommentOfProduct,deleteProductinShoppingCart } = require('../controllers/users/userController');
+const { getAllWishes,getAllUsers,addUser,addProductInShoppingCartForUser,addFriends,addWishToList,getAllFriends,getAllProductsInShoppingCart,addNewComment,getAllCommentOfUser,getAllCommentOfProduct,deleteProductinShoppingCart,acceptFriend,removeOrRejectedFriend,getAllFriendsPending } = require('../controllers/users/userController');
 
 const addNewUser = async (req,res) => {
     try {
@@ -15,21 +15,12 @@ const addNewUser = async (req,res) => {
     }; 
 };
 
-const addNewFriend = async (req,res) => {
-   try {
-    const {emailUser,emailFriend} = req.query;
-    const userAddedFriend = await addFriends(emailUser,emailFriend);
-    // if(userAddedFriend.error) throw new Error(userAddedFriend.error);
-    return res.status(200).json(userAddedFriend);
-   } catch (error) {
-    return res.status(400).json(error.message);
-   };
-};
 
 
 const addNewProductInShoppingCart = async (req,res) => {
     try {
         const {email,idProduct} = req.query;
+        if(!email || !idProduct) throw new Error('send me all data please');
         const userWithProductAdded = await addProductInShoppingCartForUser(email,idProduct);
         if(userWithProductAdded.error) throw new Error(userWithProductAdded.error);
         return res.status(200).json(userWithProductAdded);
@@ -40,6 +31,7 @@ const addNewProductInShoppingCart = async (req,res) => {
 const removeProductoInShoppingCar = async (req,res) => { 
     try {
         const {email,idProduct} = req.query;
+        if(!email || !idProduct) throw new Error('send me all data please');
         const newList = await deleteProductinShoppingCart(email,idProduct);
         if(newList.error) throw new Error(newList.error);
         return res.status(200).json(newList);
@@ -61,19 +53,62 @@ const userList = async (req,res) => {
 
 const addWish = async (req,res) => {
     const {user, product} = req.query;
+    if(!user || !product) throw new Error('send me all data please');
     const prueba = await addWishToList(user,product);
     return res.status(200).send(prueba);
 };
 const userID = async (req,res) => {
-
+    
 };
-
+const addNewFriend = async (req,res) => {
+   try {
+    const {emailUser,emailFriend} = req.query;
+    if(!emailUser || !emailFriend) throw new Error('send me all data please');
+    const userAddedFriend = await addFriends(emailUser,emailFriend);
+    // if(userAddedFriend.error) throw new Error(userAddedFriend.error);
+    return res.status(200).json(userAddedFriend);
+   } catch (error) {
+    return res.status(400).json(error.message);
+   };
+};
+const responseRequestNewFriend = async (req,res) => {
+    try {
+        const { email,emailFriend,response } = req.query;
+        if(!email || !emailFriend || !response) throw new Error('send me all data please');
+        console.log(email,emailFriend,response)
+        if(response === 'accept'){
+            const result = await acceptFriend(email,emailFriend);
+            if(result.error) throw new Error(result.error);
+            return res.status(200).json('the friend was accept');            
+        };
+        const result = await removeOrRejectedFriend(email,emailFriend,response);
+        if(result.error) throw new Error(result.error);
+        return res.status(200).json('the friend was rejected');            
+     
+          
+        // return res.status(200).json('the friend was rejected');
+    } catch (error) {
+        return res.status(400).json(error.message);
+    }
+};
 const friendsList = async (req,res) => {
     try {
         const { email } = req.query;
+        if(!email) throw new Error('send me all data please');
         const friendsList = await getAllFriends(email);
         if(friendsList.error) throw new Error(friendsList.error);
         return res.status(200).json(friendsList);
+    } catch (error) {
+        return res.status(400).json(error.message);
+    }
+};
+const friendsPendingList = async (req,res) => {
+    try {
+        const { email } = req.query;
+        if(!email) throw new Error('send me all data please');
+        const friendsPendingList = await getAllFriendsPending(email);
+        if(friendsPendingList.error) throw new Error(friendsPendingList.error);
+        return res.status(200).json(friendsPendingList);
     } catch (error) {
         return res.status(400).json(error.message);
     }
@@ -82,6 +117,7 @@ const friendsList = async (req,res) => {
 const productsListShoppingCart = async (req,res)=> {
     try {
         const { email } = req.query;
+        if(!email) throw new Error('send me all data please');
         const productList = await getAllProductsInShoppingCart(email);
         if(productList.error) throw new Error(productList.error);
         return res.status(200).json(productList);
@@ -93,6 +129,7 @@ const productsListShoppingCart = async (req,res)=> {
 const wishesList = async (req,res) => {
     try {
         const { email } = req.query;
+        if(!email) throw new Error('send me all data please');
         const productList = await getAllWishes(email);
         if(productList.error) throw new Error(productList.error);
         return res.status(200).json(productList);
@@ -103,8 +140,9 @@ const wishesList = async (req,res) => {
 
 const addComment = async (req,res) => {
     try {
-        const { email,comment,product } = req.query;
-        const newComment = await addNewComment(email,comment,product );
+        const { email,comment,productId } = req.body;
+        if(!email || !comment || !productId) throw new Error('email or comment or productId is missing. Send data correctly please...');
+        const newComment = await addNewComment(email,comment,productId );
         return res.status(200).json(newComment);
     } catch (error) {
         return res.status(400).json(error.message);
@@ -115,6 +153,7 @@ const addComment = async (req,res) => {
 const commentListOfUser = async (req,res) => {
     try {
         const { email } = req.query;
+        if(!email) throw new Error('send me all data please');
         const commentList = await getAllCommentOfUser(email);
         return res.status(200).json(commentList);
     } catch (error) {
@@ -126,6 +165,7 @@ const commentListOfUser = async (req,res) => {
 const commentListOfProduct= async (req,res) => {
     try {
         const { idProduct } = req.query;
+        if(!idProduct) throw new Error('send me all data please');
         const commentList = await getAllCommentOfProduct(idProduct);
         return res.status(200).json(commentList);
     } catch (error) {
@@ -133,4 +173,4 @@ const commentListOfProduct= async (req,res) => {
     };
 };
 
-module.exports = { userList,userID,addNewUser,addNewProductInShoppingCart,addNewFriend,addWish,friendsList,productsListShoppingCart,wishesList,addComment,commentListOfUser,commentListOfProduct,removeProductoInShoppingCar };
+module.exports = { userList,userID,addNewUser,addNewProductInShoppingCart,addNewFriend,addWish,friendsList,productsListShoppingCart,wishesList,addComment,commentListOfUser,commentListOfProduct,removeProductoInShoppingCar,responseRequestNewFriend,friendsPendingList };
